requirements.txt
anyio>=3.7.1,<4.0.0
click==8.1.7
dnspython==2.6.1
email_validator >=2.1.1
fastapi==0.111.0
graphql-core==3.2.3
greenlet==3.0.3
h11==0.14.0
idna==3.7
libcst==1.3.1
msgpack==1.0.8
mypy-extensions==1.0.0
pathspec==0.12.1
platformdirs==4.2.2
pydantic>=2.7.1,<3.0.0
Pygments==2.18.0
pymongo==4.7.2
python-multipart >=0.0.9
requests>=2.31.0
rich==13.7.1
six==1.16.0
sniffio==1.3.1
starlette>=0.37.2,<0.38.0
strawberry-graphql[fastapi]==0.229.1
tomli==2.0.1
typer==0.12.3
typing-extensions>=4.11.0
uvicorn==0.29.0
